its work - need to submit

+++ Expression.h
@@ -9,7 +9,7 @@ using namespace std;
 class Expression {
 
 public:
-    virtual double calculate() = 0;
+    virtual float calculate() = 0;
     virtual ~Expression() {}
 };
